+++ packages/core/content-manager/admin/src/components/CollectionTypeFormWrapper/index.js
