+++ packages/core/content-manager/admin/src/components/CollectionTypeFormWrapper/utils/getRequestUrl.js
