+++ packages/core/content-manager/admin/src/components/DragLayer/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { useDragLayer } from 'react-dnd';
-import LayoutDndProvider from '../../pages/LayoutDndProvider';
+import LayoutDndProvider from '../LayoutDndProvider';
 
 import ItemTypes from '../../utils/ItemTypes';
