+++ packages/core/content-manager/admin/src/components/EditViewDataManagerProvider/reducer.js
