+++ packages/core/content-manager/admin/src/components/EditViewDataManagerProvider/utils/cleanData.js
