+++ packages/core/content-manager/admin/src/components/EditViewDataManagerProvider/utils/schema.js
