+++ packages/core/content-manager/admin/src/components/LayoutDndProvider/index.js
