+++ packages/core/content-manager/admin/src/components/ListViewProvider/index.js
