+++ packages/core/content-manager/admin/src/components/SingleTypeFormWrapper/index.js
@@ -22,7 +22,7 @@ import {
 } from '../../sharedReducers/crudReducer/actions';
 import selectCrudReducer from '../../sharedReducers/crudReducer/selectors';
 import { getRequestUrl } from './utils';
-import buildQueryString from '../ListView/utils/buildQueryString';
+import buildQueryString from '../../pages/ListView/utils/buildQueryString';
 
 // This container is used to handle the CRUD
 const SingleTypeFormWrapper = ({ allLayoutData, children, slug }) => {
