+++ packages/core/content-manager/admin/src/components/SingleTypeFormWrapper/utils/getRequestUrl.js
