+++ packages/core/content-manager/admin/src/components/SingleTypeFormWrapper/utils/index.js
