+++ packages/core/content-manager/admin/src/components/WysiwygProvider/index.js
