+++ packages/core/content-manager/admin/src/hooks/index.js
@@ -4,3 +4,4 @@ export { default as useFindRedirectionLink } from './useFindRedirectionLink';
 export { default as useLayoutDnd } from './useLayoutDnd';
 export { default as useListView } from './useListView';
 export { default as useWysiwyg } from './useWysiwyg';
+export { default as useSyncRbac } from './useSyncRbac';
