+++ packages/core/content-manager/admin/src/hooks/useSyncRbac/index.js
