+++ packages/core/content-manager/admin/src/pages/EditView/index.js
@@ -18,9 +18,9 @@ import FormWrapper from '../../components/FormWrapper';
 import FieldComponent from '../../components/FieldComponent';
 import Inputs from '../../components/Inputs';
 import SelectWrapper from '../../components/SelectWrapper';
-import CollectionTypeFormWrapper from '../CollectionTypeFormWrapper';
-import EditViewDataManagerProvider from '../EditViewDataManagerProvider';
-import SingleTypeFormWrapper from '../SingleTypeFormWrapper';
+import CollectionTypeFormWrapper from '../../components/CollectionTypeFormWrapper';
+import EditViewDataManagerProvider from '../../components/EditViewDataManagerProvider';
+import SingleTypeFormWrapper from '../../components/SingleTypeFormWrapper';
 import Header from './Header';
 import { createAttributesLayout, getFieldsActionMatchingPermissions } from './utils';
 import { LinkWrapper, SubWrapper } from './components';
