+++ packages/core/content-manager/admin/src/pages/ListViewLayoutManager/index.js
@@ -3,9 +3,8 @@ import { useDispatch } from 'react-redux';
 import PropTypes from 'prop-types';
 import { useHistory } from 'react-router-dom';
 import { useQueryParams } from '@strapi/helper-plugin';
-import { useFindRedirectionLink } from '../../hooks';
+import { useFindRedirectionLink, useSyncRbac } from '../../hooks';
 import { resetProps, setLayout } from '../ListView/actions';
-import useSyncRbac from '../RBACManager/useSyncRbac';
 import Permissions from './Permissions';
 
 const ListViewLayout = ({ layout, ...props }) => {
