+++ packages/core/content-manager/admin/src/reducers.js
@@ -1,7 +1,7 @@
 import mainReducer from './pages/Main/reducer';
 import editViewLayoutManagerReducer from './pages/EditViewLayoutManager/reducer';
 import listViewReducer from './pages/ListView/reducer';
-import rbacManagerReducer from './pages/RBACManager/reducer';
+import rbacManagerReducer from './hooks/useSyncRbac/reducer';
 import editViewCrudReducer from './sharedReducers/crudReducer/reducer';
 import pluginId from './pluginId';
